Moved module list into separate file. Added module list locations endpoint. Updated readme.
